+++ test/runtests.jl
@@ -16,9 +16,12 @@ freqs = linspace(0.01, 3, nfreqs)
 # Randomize frequency grid
 freqs += step(freqs)*rand(nfreqs)
 freqs = collect(freqs)
-# Use "freqpower" just to call that function and increase code coverage.
-# "autofrequency" function is tested below.
-@test_approx_eq_eps freqpower(lombscargle(t, s, frequencies=freqs, fit_mean=false))[2] freqpower(lombscargle(t, s, frequencies=freqs, fit_mean=true))[2] 5e-3
+# Use "freqpower" and "periodpower" just to call that function and increase code
+# coverage.  "autofrequency" function is tested below.
+pgram1 = lombscargle(t, s, frequencies=freqs, fit_mean=false)
+pgram2 = lombscargle(t, s, frequencies=freqs, fit_mean=true)
+@test_approx_eq_eps freqpower(pgram1)[2]   freqpower(pgram2)[2] 5e-3
+@test_approx_eq_eps periodpower(pgram1)[2] periodpower(pgram2)[2] 5e-3
 
 # Simple signal, without any randomness
 t = collect(linspace(0.01, 10pi, ntimes))
@@ -72,11 +75,15 @@ pgram6 = lombscargle(t, s, maximum_frequency=30, fast=false)
 # Test findmaxfreq and findmaxpower
 @test_approx_eq findmaxfreq(pgram1)        [31.997145470342]
 @test_approx_eq findmaxfreq(pgram1, 0.965) [0.15602150741832602,31.685102455505348,31.997145470342,63.52622641842902,63.838269433265665]
+@test_approx_eq findmaxperiod(pgram1) 1./findmaxfreq(pgram1)
+@test_approx_eq findmaxperiod(pgram1, 0.965) 1./findmaxfreq(pgram1, 0.965)
 t = linspace(0.01, 10pi, 1001)
 s = sinpi(2t) + cospi(4t)
 p = lombscargle(t, s, maximum_frequency=4)
 @test_approx_eq findmaxfreq(p, [0.9, 1.1]) 1.0029954048320957
 @test_approx_eq findmaxfreq(p, [1.9, 2.1]) 2.002806697267899
+@test_approx_eq findmaxperiod(p, [1/0.9, 1/1.1]) 1./findmaxfreq(p, [0.9, 1.1])
+@test_approx_eq findmaxperiod(p, [1/1.9, 1/2.1]) 1./findmaxfreq(p, [1.9, 2.1])
 @test_approx_eq findmaxpower(pgram1) 0.9695017551608017
 
 # Test autofrequency function